backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0
xgboost>=2.0.0
lightgbm>=4.3.0
scikit-learn>=1.4.0
keras==3.13.2
tensorflow-cpu>=2.20.0
numpy>=1.26.0
opencv-python-headless>=4.9.0
pillow>=10.3.0
python-dotenv>=1.0.0
